# Improve ZVOL sync write performance by using a taskq