@ahrens 2020-03-30
Relates to previous changes https://github.com/openzfs/zfs/pull/3720,   https://github.com/openzfs/zfs/pull/5824  and https://github.com/openzfs/zfs/pull/6477