@filip-paczynski 2020-04-02
This is great news! 
Is applying the changes from #3720, #5824 and #6477 over 0.8.3 will be enough to test this?