@steven-b89 2020-05-12
@ahrens did you test the zvol > iscsi on a separate host or did you run ZFS inside Hyper-V ?
I am very interested in this topic to improve sync performance, however I compiled the master from here with no difference at all in sync performance over iscsi.