@ahrens 2020-11-11
This PR is in OpenZFS 2.0 (all RC's).  It isn't part of 0.8.x.